Split instance.ts into multiple files
